Rombak besar2an tenant

+++ app/Http/Controllers/TenantController.php
@@ -10,6 +10,7 @@ use App\Models\MsTenant;
 use App\Models\MsTenantType;
 use App\Models\User;
 use App\Models\MsUnitOwner;
+use App\Models\TrContract;
 use Validator;
 
 class TenantController extends Controller
@@ -80,6 +81,7 @@ class TenantController extends Controller
                 $temp['tent_name'] = $value->tent_name;
                 $temp['tenan_isppn'] = $value->tenan_isppn;
                 $temp['tenan_ispkp'] = $value->tenan_ispkp;
+                $temp['action'] = '<a href="#" data-id="'.$value->id.'" class="addUnit" title="Add Unit"><i class="fa fa-home" aria-hidden="true"></i> Add Unit</a>';
                 $result['rows'][] = $temp;
             }
             return response()->json($result);
@@ -142,6 +144,13 @@ class TenantController extends Controller
             $input['created_by'] = Auth::id();
             $input['updated_by'] = Auth::id();
 
+            if(isset($input['unit_id'])){
+                // cek unit
+                $cekunit = MsUnitOwner::where('unit_id',$input['unit_id'])->first();
+                if($cekunit){
+                    return response()->json(['errorMsg' => 'Unit is already owned by others']);       
+                }
+            }
             $tenant = MsTenant::create($input);
 
             if(isset($input['unit_id'])){
@@ -156,8 +165,8 @@ class TenantController extends Controller
     }
 
     public function update(Request $request){
-        // try{
-        // var_dump($request->all());
+        try{
+        // var_dump($request->all()); die();
             $id = $request->id;
             $messages = [
                 'tenan_code.unique' => 'Tenant Code must be unique',
@@ -180,15 +189,30 @@ class TenantController extends Controller
             $input = $request->all();
             $input['updated_by'] = Auth::id();
             MsTenant::find($id)->update($input);
-            return MsTenant::find($id);
-        // }catch(\Exception $e){
-        //     return response()->json(['errorMsg' => $e->getMessage()]);
-        // } 
+
+            $tenant = MsTenant::find($id);
+
+            // $current_unit_id = @$input['current_unit_id'];
+            // if(isset($input['unit_id']) && ($current_unit_id != $input['unit_id'])){
+            //     if(!empty($current_unit_id)) MsUnitOwner::where('unit_id',$input['unit_id'])->where('tenan_id',$tenant->id)->delete();
+            //     MsUnitOwner::create(['unit_id'=>$input['unit_id'], 'tenan_id'=>$tenant->id]);
+            // }
+
+            return $tenant;
+        }catch(\Exception $e){
+            return response()->json(['errorMsg' => $e->getMessage()]);
+        } 
     }
 
     public function delete(Request $request){
         try{
             $id = $request->id;
+            // cek kontrak berjalan
+            $now = date('Y-m-d');
+            $cekKontrak = TrContract::whereNull('contr_terminate_date')->where('contr_startdate','<=',$now)->where('contr_enddate','>=',$now)
+                        ->where('tenan_id',$id)->first();
+            if($cekKontrak) return response()->json(['errorMsg' => 'There are any contract running for this tenant, You can remove this after contract is not available for this tenant']);
+
             MsTenant::destroy($id);
             return response()->json(['success'=>true]);
         }catch(\Exception $e){
@@ -214,5 +238,46 @@ class TenantController extends Controller
             array_push($result['results'], $temp);
         }
         return json_encode($result);
+    }
+
+    public function modaldt(Request $request){
+        $id = $request->id;
+        $fetch = MsTenant::select('ms_tenant.*','ms_tenant_type.tent_name')
+                        ->join('ms_tenant_type','ms_tenant.tent_id','=','ms_tenant_type.id')
+                        ->where('ms_tenant.id',$id)->first();
+        $units = MsUnitOwner::select('ms_unit_owner.unit_id','ms_unit.unit_code','ms_unit.unit_name','ms_unit.unit_sqrt','ms_virtual_account.viracc_no')
+                            ->join('ms_unit','ms_unit_owner.unit_id','=','ms_unit.id')
+                            ->join('ms_virtual_account','ms_unit.unit_virtual_accn','=','ms_virtual_account.id')
+                            ->where('tenan_id',$id)->get();
+        return view('modal.tenantdetail', ['id'=>$id,'fetch' => $fetch, 'units'=>$units]);
+    }
+
+    public function deleteunit(Request $request){
+        try{
+            $unitid = $request->unitid;
+            $tenanid = $request->tenanid;
+            MsUnitOwner::where('unit_id',$unitid)->where('tenan_id',$tenanid)->delete();
+            return response()->json(['success'=>true]);
+        }catch(\Exception $e){
+            return response()->json(['errorMsg' => $e->getMessage()]);
+        } 
+    }
+
+    public function addunit(Request $request){
+        try{
+            $unitid = $request->unitid;
+            $tenanid = $request->tenanid;
+            $date = $request->date;
+            
+            $cekUnit = MsUnitOwner::where('unit_id',$unitid)->where('tenan_id',$tenanid)->first();
+            if($cekunit) return response()->json(['errorMsg' => 'Unit is already owned']);
+            $cekUnit = MsUnitOwner::where('unit_id',$unitid)->first();
+            if($cekunit) return response()->json(['errorMsg' => 'Unit is already owned by others']);
+
+            MsUnitOwner::create(['unit_id'=>$unitid,'tenan_id'=>$tenanid, 'unitow_start_date'=>$date]);
+            return response()->json(['success'=>true]);
+        }catch(\Exception $e){
+            return response()->json(['errorMsg' => $e->getMessage()]);
+        } 
     }
 }
